requirements/base.txt
Django==1.9.7
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
django-appconf==1.0.1
six==1.10.0
django-tastypie==0.13.3